style: initial revamp
Revamping the UI to make it more easy to interact and use.
